#3343: Refactor cb init
Move cb init code into circular_buffer.h
Use the same init code for brisc/ncrisc/trisc
Pull cq init out of brisk and into command_queue
