+++ tt_metal/src/firmware/riscv/common/circular_buffer.h
@@ -6,6 +6,8 @@
 
 #pragma once
 
+#include "hostdevcommon/common_runtime_address_map.h"
+#include "risc_attribs.h"
 
 struct CQReadInterface {
     uint32_t fifo_size;
@@ -35,3 +37,37 @@ struct CBInterface {
     // used by packer for in-order packing
     uint32_t fifo_wr_tile_ptr;
 };
+
+extern CBInterface cb_interface[NUM_CIRCULAR_BUFFERS];
+
+// NCRISC and BRISC setup read and write
+// TRISC sets up read or write
+inline void setup_cb_read_write_interfaces(bool read, bool write) {
+    volatile tt_l1_ptr uint32_t* circular_buffer_config_addr = (volatile tt_l1_ptr uint32_t*)(CIRCULAR_BUFFER_CONFIG_BASE);
+
+    for (uint32_t cb_id = 0; cb_id < NUM_CIRCULAR_BUFFERS; cb_id++) {
+
+        // NOTE: fifo_addr, fifo_size and fifo_limit in 16B words!
+        uint32_t fifo_addr = circular_buffer_config_addr[0];
+        uint32_t fifo_size = circular_buffer_config_addr[1];
+        uint32_t fifo_num_pages = circular_buffer_config_addr[2];
+        uint32_t fifo_page_size = circular_buffer_config_addr[3];
+        uint32_t fifo_limit = fifo_addr + fifo_size;
+
+        cb_interface[cb_id].fifo_limit = fifo_limit;  // to check if we need to wrap
+        if (write) {
+            cb_interface[cb_id].fifo_wr_ptr = fifo_addr;
+        }
+        if (read) {
+            cb_interface[cb_id].fifo_rd_ptr = fifo_addr;
+        }
+        cb_interface[cb_id].fifo_size = fifo_size;
+        cb_interface[cb_id].tiles_acked_received_init = 0;
+        if (write) {
+            cb_interface[cb_id].fifo_num_pages = fifo_num_pages;
+        }
+        cb_interface[cb_id].fifo_page_size = fifo_page_size;
+
+        circular_buffer_config_addr += UINT32_WORDS_PER_CIRCULAR_BUFFER_CONFIG;
+    }
+}
