+++ tt_metal/src/firmware/riscv/targets/brisc/brisck.cc
@@ -26,7 +26,6 @@
 #include "debug_print.h"
 
 CBInterface cb_interface[NUM_CIRCULAR_BUFFERS];
-CQReadInterface cq_read_interface;
 
 uint8_t noc_index = NOC_INDEX;
 
@@ -34,11 +33,7 @@ void kernel_launch() {
 
     firmware_kernel_common_init((void tt_l1_ptr *)MEM_BRISC_INIT_LOCAL_L1_BASE);
 
-#if defined(IS_DISPATCH_KERNEL)
-    setup_cq_read_write_interface();
-#else
-    setup_cb_read_write_interfaces();                // done by both BRISC / NCRISC
-#endif
+    setup_cb_read_write_interfaces(true, true);
 
     noc_local_state_init(noc_index);
