+++ tt_metal/src/firmware/riscv/targets/ncrisc/ncrisck.cc
@@ -17,7 +17,6 @@
 #include "kernel.cpp"
 
 CBInterface cb_interface[NUM_CIRCULAR_BUFFERS];
-CQReadInterface cq_read_interface;
 
 uint8_t noc_index = NOC_INDEX;
 
@@ -29,7 +28,7 @@ void kernel_launch() {
 
     firmware_kernel_common_init((void tt_l1_ptr *)MEM_NCRISC_INIT_LOCAL_L1_BASE);
 
-    setup_cb_read_write_interfaces();
+    setup_cb_read_write_interfaces(true, true);
 
     noc_local_state_init(noc_index);
